requirements.txt
Flask
h5py
names-dataset==1.9.1
numpy==1.19.2
tensorflow-cpu
gunicorn
keras
